Refactored card layout logic to be cleaner
