+++ index.d.ts
@@ -246,9 +246,11 @@ interface KickerType {
     showSlash?: boolean;
 }
 
+type ImagePositionType = 'left' | 'top' | 'right';
+
 type CardImageType = {
     element: ImageBlockElement;
-    position?: 'left' | 'top' | 'right';
+    position?: ImagePositionType;
     size?: ImageSizeType;
 };
